Rename ImageryViewLayout to ImageryView Fix bug for View Providers to check for navigated object Added resize observer and removed setInterval Refactored isNavigatedObject utility method to Application Router
