+++ src/plugins/imagery/components/ImageryView.vue
