+++ src/plugins/imagery/components/imagery-view.scss
